+++ tracker/query/tvshows.py
@@ -1,0 +1,73 @@
+from typing import Literal
+
+from tracker.query.base import TVMovieDB
+from tracker.query.model import Show, TVShowLength
+
+
+class TelevisionDB(TVMovieDB):
+    def get_genres(self) -> dict:
+        response = self.request("/genre/tv/list", params={"language": "en-US"})
+        return response
+
+    def search(self, query: str, include_adult: bool = False) -> dict:
+        response = self.request(
+            "/search/tv",
+            params={
+                "query": query,
+                "include_adult": include_adult,
+                "language": "en-US",
+                "page": "1",
+            },
+        )
+        return response
+
+    def get_trending(self, timeframe: Literal["day", "week"]) -> dict:
+        response = self.request(
+            f"/trending/tv/{timeframe}",
+            params={"language": "en-US"},
+        )
+        return response
+
+    def get_series(self, series_id: int):
+        response = self.request(f"/tv/{series_id}", params={"language": "en-US"})
+        return response
+
+
+class TVShowAdapter:
+    def __init__(self, tv_db: TelevisionDB):
+        self.tv_db = tv_db
+        self.genres_by_id = {
+            genre["id"]: genre["name"] for genre in tv_db.get_genres()["genres"]
+        }
+
+    def hydrate_tvshow(self, show_reference: dict):
+        series = self.tv_db.get_series(show_reference["id"])
+        return Show(
+            name=show_reference["name"],
+            description=show_reference["overview"],
+            genres=[self.genres_by_id[genre] for genre in show_reference["genre_ids"]],
+            length=TVShowLength(
+                seasons=series["number_of_seasons"],
+                episodes_per_season={
+                    item["season_number"]: item["episode_count"]
+                    for item in series["seasons"]
+                },
+            ),
+            networks=[network["name"] for network in series["networks"]],
+        )
+
+    def search(self, query: str) -> list[Show]:
+        page = self.tv_db.search(query)
+        items = page["results"]
+        results = []
+        for item in items:
+            results.append(self.hydrate_tvshow(item))
+        return results
+
+    def trending(self, timeframe: Literal["day", "week"]) -> list[Show]:
+        page = self.tv_db.get_trending(timeframe)
+        items = page["results"]
+        results = []
+        for item in items:
+            results.append(self.hydrate_tvshow(item))
+        return results
